requirements.txt
# Nexus RAG — Python Dependencies
# Generated for Python 3.12
# Install with: pip install -r requirements.txt

# ── Web Framework ──────────────────────────
flask>=3.0.0
flask-cors>=4.0.0

# ── LangChain Core ────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# ── LLM Provider: Groq ────────────────────
langchain-groq>=0.2.0

# ── Embeddings ────────────────────────────
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0

# ── Vector Store ──────────────────────────
faiss-cpu>=1.8.0

# ── PDF Loading ───────────────────────────
pypdf>=4.0.0

# ── Utilities ─────────────────────────────
python-dotenv>=1.0.0
werkzeug>=3.0.0
